requirements.txt
flake8==4.0.1
fuzzywuzzy==0.18.0
python-dotenv==0.19.2
python-Levenshtein==0.12.2
python-telegram-bot==13.11
redis==4.2.1
requests==2.27.1
vk-api==11.9.8
